Add generic Result class and error handling test

+++ app/src/main/java/com/xpense/android/data/TransactionRepositoryImpl.kt
@@ -1,6 +1,8 @@
 package com.xpense.android.data
 
 import androidx.lifecycle.LiveData
+import com.xpense.android.data.Result.Error
+import com.xpense.android.data.Result.Success
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
 
@@ -10,7 +12,7 @@ class TransactionRepositoryImpl constructor(
     private val ioDispatcher: CoroutineDispatcher = Dispatchers.IO
 ) : TransactionRepository {
 
-    override fun observeTransactions(): LiveData<List<Transaction>> =
+    override fun observeTransactions(): LiveData<Result<List<Transaction>>> =
         transactionDataSourceLocal.observeTransactions()
 
     override suspend fun insertTransaction(transaction: Transaction) =
@@ -19,20 +21,26 @@ class TransactionRepositoryImpl constructor(
     override suspend fun getTransaction(transactionId: Long) =
         transactionDataSourceLocal.getTransaction(transactionId)
 
-    override suspend fun getTransactions(): List<Transaction> =
+    override suspend fun getTransactions(): Result<List<Transaction>> =
         transactionDataSourceLocal.getTransactions()
 
     override suspend fun updateTransaction(transaction: Transaction) =
         transactionDataSourceLocal.updateTransaction(transaction)
 
-    override suspend fun refreshTransactions() {
+    override suspend fun refreshTransactions() =
+        updateTransactionsFromRemoteDataSource()
+
+    private suspend fun updateTransactionsFromRemoteDataSource() {
         val remoteTransactions = transactionDataSourceRemote.getTransactions()
 
-        // Real apps might want to do a proper sync
-        transactionDataSourceLocal.deleteAllTransactions()
-        remoteTransactions.forEach { transaction ->
-            transactionDataSourceLocal.saveTransaction(transaction)
+        if (remoteTransactions is Success) {
+            // Real apps might want to do a proper sync
+            transactionDataSourceLocal.deleteAllTransactions()
+            remoteTransactions.data.forEach { transaction ->
+                transactionDataSourceLocal.saveTransaction(transaction)
+            }
+        } else if (remoteTransactions is Error) {
+            throw remoteTransactions.exception
         }
     }
-
 }
